requirements.txt
robotframework==7.1.1
robotframework-seleniumlibrary==0.9.7
Faker==4.0.0
pyyaml=6.0.2